added

+++ tibco-webservice-batch/spring-boot/SpringBootRestJpa/src/main/java/com/cts/service/InsuranceService.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.cts.dao.InsuranceRepository;
+import com.cts.exception.InsuranceException;
 import com.cts.model.Insurance;
 
 import lombok.Data;
@@ -19,32 +20,48 @@ public class InsuranceService {
 	private InsuranceRepository repository;
 
 	public Insurance addInsurance(Insurance insurance) {
+		Optional<Insurance> byId = repository.findById(insurance.getId());
+		if (byId.isPresent())
+			throw new InsuranceException("id already  present");
+
 		return repository.save(insurance);
 	}
 
 	public Insurance updateInsurance(Insurance insurance) {
+
+		Optional<Insurance> byId = repository.findById(insurance.getId());
+		if (byId.isEmpty())
+			throw new InsuranceException("id not  present");
+
 		return repository.save(insurance);
 	}
 
 	public List<Insurance> getAllInsurance() {
+		if (repository.findAll().isEmpty())
+			throw new InsuranceException("insurance list is empty");
 		return repository.findAll();
 	}
 
 	public Insurance deleteInsurance(int id) {
-		Insurance insurance = null;
+
 		Optional<Insurance> byId = repository.findById(id);
-		if (byId.isPresent())
-			insurance = byId.get();
+		if (byId.isEmpty())
+			throw new InsuranceException("id not  present");
+
+		Insurance insurance = null;
+		insurance = byId.get();
 
 		repository.deleteById(id);
 		return insurance;
 	}
 
 	public Insurance searchInsurance(int id) {
+
 		Insurance insurance = null;
 		Optional<Insurance> byId = repository.findById(id);
-		if (byId.isPresent())
-			insurance = byId.get();
+		if (byId.isEmpty())
+			throw new InsuranceException("id nou found ");
+		insurance = byId.get();
 
 		return insurance;
 	}
